Feat: move redesigned timestamp on top of card
