+++ dotcom-rendering/src/components/MediaMeta.tsx
@@ -1,5 +1,4 @@
 import { css } from '@emotion/react';
-import { textSans } from '@guardian/source-foundations';
 import { SvgAudio, SvgCamera } from '@guardian/source-react-components';
 import { decidePalette } from '../lib/decidePalette';
 import type { DCRContainerPalette } from '../types/front';
@@ -34,38 +33,11 @@ const iconWrapperStyles = (palette: Palette, hasKicker: boolean) => css`
 	}
 `;
 
-const durationStyles = (palette: Palette, hasKicker: boolean) => css`
-	color: ${hasKicker ? palette.text.cardKicker : palette.text.cardFooter};
-	${textSans.xxsmall({ fontWeight: `bold` })}
-`;
-
 const wrapperStyles = css`
 	display: flex;
 	align-items: center;
 	margin-top: 4px;
 `;
-
-export function secondsToDuration(secs?: number): string {
-	if (typeof secs === `undefined` || secs === 0) {
-		return ``;
-	}
-	const seconds = Number(secs);
-	const h = Math.floor(seconds / 3600);
-	const m = Math.floor((seconds % 3600) / 60);
-	const s = Math.floor((seconds % 3600) % 60);
-
-	const duration = [];
-	if (h > 0) {
-		duration.push(h);
-	}
-	if (h > 0 && m < 10) duration.push(`0${m}`); // e.g 1:01:11
-	else duration.push(m); // supports 0:59
-	if (s > 0) {
-		if (s < 10) duration.push(`0${s}`);
-		else duration.push(s);
-	}
-	return duration.join(':');
-}
 
 const Icon = ({ mediaType }: { mediaType: MediaType }) => {
 	switch (mediaType) {
@@ -95,23 +67,8 @@ const MediaIcon = ({
 	);
 };
 
-const MediaDuration = ({
-	mediaDuration,
-	palette,
-	hasKicker,
-}: {
-	mediaDuration: number;
-	palette: Palette;
-	hasKicker: boolean;
-}) => (
-	<p css={durationStyles(palette, hasKicker)}>
-		{secondsToDuration(mediaDuration)}
-	</p>
-);
-
 export const MediaMeta = ({
 	mediaType,
-	mediaDuration,
 	format,
 	containerPalette,
 	hasKicker,
@@ -124,14 +81,6 @@ export const MediaMeta = ({
 				palette={palette}
 				hasKicker={hasKicker}
 			/>
-			&nbsp;
-			{mediaDuration !== undefined && mediaDuration > 0 && (
-				<MediaDuration
-					mediaDuration={mediaDuration}
-					palette={palette}
-					hasKicker={hasKicker}
-				/>
-			)}
 		</div>
 	);
 };
